Refactor 'filter_note_path()' so that it returns a 'Result'.
In particular, it returns the warnings that it used to simply print.
Add 'UnPushedPathWarning'.
Add 'NotAnkiNoteWarning'.
Add `test_push_honours_ignore_patterns()`.
Change the return type of 'diff_repos()' so that it returns a
heterogeneous list of either 'Delta's or 'Warning's.
Print out all warnings in 'push_deltas()'.
Remove unused in port in 'ki/types.py'.