broken: Add '_pull()' function.
This is a first attempt at fixing a bug where 'push()' thinks it is up
to date when it is not.
Add a call to '_pull()' on the stage repo within 'push()'.
Add many comments.
Lint files partially.
Failure of 'test_push_is_nontrivial_when_pulled_changes_are_reverted()'
is due to the '_pull()' call on a 'stage_kirepo' whose '.ki/' directory
contents are out-of-sync with its git history. In particular, the LAST
PUSH commit ref may reference a commit that does not exist in the
history.