Add logic to remove submodule-associated directories from remote repo.
Above must be done before pulling in order to avoid overwriting
submodules.
Submodules are pulled separately by first unsubmoduling a copy of a
checkout of the local trunk at the ref of the last successful PUSH, and
then diffing this from HEAD~FETCH_HEAD. Then patches are generated for
each submodule, and the patches are applied within the submodules
instead, where we commit the changes.