Remove a `git_pull()` call from `_pull()`.
This commit replaces a `git pull --allow-unrelated-histories -X theirs`
with the same procedure used in `push()`. We graft the `.git/` directory
onto the working tree of a fresh clone of the database to construct the
staging repo. Then we pull the stage into the main ki repository.

This fixes a bug where subdecks were duplicated during pulls in the case
where a deck was moved in Anki.