Refactor `test_push_doesnt_fail_after_pull()` to use `NOTE_7`.
Caution: this looks like a regression. This reflects the use of a
`.git/` graft in `_pull()` in-place of the previously used internal
`git_pull()` call, allowing unrelated histories. This has the effect of
making renames more commmon (correctly) and results in more
fast-forwards during the final, external `git_pull()` call.

We use the terms 'internal' and 'external' to refer to `git pull` calls
that either should always run successfully, or could possibly result in
a merge conflict that the user must deal with, respectively.