Add support for creation of git symlinks on Windows during `clone()`.
This commit adds functionality to support POSIX-style symlinks on
Windows. In particular, we create the same sort of pseudo-symlinks that
git creates when cloning repositories on Windows systems with actual
symlinks disabled. They are simply ordinary files with their `git` file
mode set to `120000`, and whose contents are simply the path to the
target.

* Make `filemode()` take all extant file types.
* Make `filemode()` return `-1` when file is untracked.
* Remove `git.reset()` calls at end-of-`clone()` that only existed to
  prevent Windows behavior tests from breaking stuff on POSIX. We now
  simply only make those assertions on Windows.
* Add some debugging statements to `test_push_writes_media()`.
* Unskip all integration tests.
* Add unit test `test_copy_repo_preserves_git_symlink_file_modes()`.