Fix bug where special chars in filenames from `guid`s crash on `win32`.
This commit fixes a bug in the implementation of `get_note_path()` where
the failsafe slugs that are autogenerated from the `guid`s of each note
are allowed to contain characters that are invalid path chars on
Windows. This is a vestigial feature of when we used to do the same
thing to generate slugs when we were using `nid`s instead of `guid`s.
Additionally, we spruce up the slug generation logic in
`get_note_path()` to use more of the note's payload. We now take content
from all fields when trying to generate a nice filename, and when that
fials, we hash the guid of the note file using `blake2`, and construct a
nice-looking filename `<notetype>--<hash>--<datetime>.md` that is a vast
improvement over the previous format (`<nid>.md`). The tests are
refactored accordingly, we add a new test for for special characters in
paths on Windows.

* Make `backup()` return an integer status code.
* Bump `MAX_FILENAME_LEN` from `40` to `60`.
* Pin `pytest` version to `7.1.2`.
* Add
  `test_write_deck_node_cards_does_not_fail_due_to_special_characters_in_paths_on_windows()`.
* Remove `logger.debug()` calls from several tests.
* Refactor `test_write_repository_displays_missing_media_warnings()` to
  avoid use of `capfd` fixture, which sometimes crashes pytest on
  Windows.
* Add type hints for all uses of `tmp_path: Path` fixture.
* Replace use of `capfd` fixture in
  `test_backup_is_no_op_when_backup_already_exists()` with a mock.
* Remove use of `capfd` from
  `test_diff2_shows_no_changes_when_no_changes_have_been_made()`.
* Refactor `test_get_note_path_produces_nonempty_filenames()` to reflect
  new implementation.
* Add helper function `get_basic_colnote_with_fields()` for tests.