Rewrite `diff2()` from the ground-up.
This commit is a full refactor of the diff procedure. We remove a lot of
unused `Warning` subtypes, and use `map()` calls in place of for-loops.
The processing of each individual diff is now done by `mungediff()`, and
as a result, the line count of `diff2()` has been reduced dramatically.

* Replace `get_note_warnings()` with `is_ignorable()`.
* Change the way several `ki` constants are defined to make things
  cleaner.
* Make `diff2()` return an `Iterable` instead of a `List`.
* Modify tests accordingly.