Add profiling code to stateful SQL parser tests.
This commit adds use of a prototype `checkpoint_anki2()` function
that saves a copy of the database dump to a temp directory. This may not
be working correctly yet, since we do not see any nontrivial parses
during tests. It is also notable that the tests did not achieve the
speedup that was hoped for, since the data generation itself is quite
expensive due to the complexity of the rules.