Add beginnings of implementation of `clone`.
This commit adds a basic implementation of parsing the target collection
file and target directory for the `clone` op. We also add dependencies
on `path`, `path-io`, `pureMD5`, and `text`.

Most file/directory actions will be done with `path-io`, which is very
nice. Several types are added for working with files and directories
which are meant to be e.g. extant or empty, as well as the functions for
parsing them and converting between them.

* Add Anki database schema reference.
* Add `*.swp` files to `.gitignore`.