Add `writeRepo` function where we make DB queries.
This commit adds a dependency on `microaeson` for dumping the config
file to `.ki/config`. Note that we used `ConfigParser` in the python
version, which wrote a TOML-like INI format. As an aside, it may be a
good idea to drop the use of JSON here, because currently the `config`
file stores only the path the collection. Better to call it `colpath`
and just read and write its contents directly. Much simpler, and doesn't
leave hooks lying around.

* Add `SQLNote` and `SQLNotetype` types and `FromRow` instances for
  parsing raw data from the Anki DB.
* Add dependency on `sqlite-simple`.
* Add `writeInitialCommit` function where we call `writeRepo` and then
  perform an initial commit of the repository.