Add functions for preprocessing fields.
This commit adds dependencies on:
        * containers
        * slugify
        * text-icu
        * text-regex-replace
        * protobuf

We also add an implementation of the `plainToHtml` function for note
fields, and a function `getColNote` for converting `SQLNote`s to
`ColNote`s.

* The usage of `Fields fs` is incorrect on line 153, and breaks this
  build. Thus, this commit should be rebased and amended.
* Similarly, the `containers` dependency should be added where we first
  use it in earlier commits.
* The constructor of `MdNote` is wrong (typo) in the previous commit,
  and should be fixed.
* Add dependency on `protobuf` for decoding config messages.

Add `Field`, `Template`, and `Model` types.

This commit also implements `getColNote` which converts a raw `SQLNote`
into a `ColNote`. The purpose of this is to transform our notes into a
form that is amenable to markdown serialization. We also add raw `SQL*`
types for models, fields, and templates, and rename notetypes to models.
A lot of logic is added for translating between maps of things indexed
by mids and ords.