Fix bug where beartype exception was raised on empty GUIDs.
This commit fixes issue #113. Our test coverage did not extend to the
`else` branch of the `guid` definition in `parse_note()`, and thus this
type error was not visible. The fix is simple (we simply grab the values
of the `flatnote.fields` map and cast to a list), and we add a test that
correctly surfaces the bug.

* Bump version to `0.0.8a`.