Add draft of media filename parser and bindings to `libtidy`.
This commit is progress towards a working implementation of `mkPayload`,
which constructs the markdown note contents from a value of type
`ColNote`. This requires us to escape media filenames in HTML tags. The
Python3 version of this project used Anki's rust backend to do this, but
we implement those regex routines from scratch in Attoparsec.

The source of the relevant regexes as well as some explanation is
provided in `UNESCAPE`.

Furthermore, we add functionality to use `html-tidy` to tidy the HTML
contained in note fields. This requires us to add bindings to `libtidy`,
the C scaffolding for which is contained in `tidy.c`.

We also change the structure of the `Fields` type, which before was a
map from names to content, but is now a list of values of type `Field`,
which is a new type containing the Ord, Name, and Content of a field.
The type that used to be called `Field` is now called `FieldDef`, which
is more intuitive.

* Add dependencies on `attoparsec`, `replace-attoparsec`.
* Use `ForeignFunctionInterface` language extension.