Add fields for building `tidylib` in `package.yaml`.
This commit adds more progress towards getting the C FFI import working
so that we can tidy HTML. We set the include and `c-sources` fields in
`package.yaml`, which still results in a linking error, but one with a
bit more detail and which makes sense. In particular, it complains about
undefined references to the tidy functions called in `htidy.h` instead
of complaining about not being able to find the `tidy` function at all.

We also remove dependencies on `text-icu` and friends, since they
require `libicu`, which is dumb.

We also add an example `makefile`.