Check python version is `3.9+` in install command
This commit replaces the install command with a bash one-liner that
dumps the python version on the `$PATH`, checks that one of the
supported versions is installed, and uses `python3 -m pip` to install it
from the `main` branch, dumping an error message if the version check
failed.

This should fix #114, in which the `pip` and `pip3` versions were
different, so `python3 --version` yielded a correct version, but the
`pip install ...` command still failed.

* Specify Python 3.9+ in the install instructions