Add working implementation of `clone`
This commit completes the `writeInitialCommit` function, introducing the
first version of the clone op that successfully builds. The major
changes are a function `gitIsDirty` for checking a repository for
unstaged or uncommitted changes, and an `appendHash` function to record
the hash of the collection associated with each operation.

We also add a bunch of attoparsec parsers for parsing HTML media tags,
along with a bare-bones `main` function with command-line argument
parsing.

Additionally, we combine several initial bug fixes for the `clone`
workflow into a single diff.

At a high-level, we:
* Eliminate use of `fieldsAndTemplatesByOrd`
* Fix bugs in core logic of `mapFieldsByMid` and `mapTemplatesByMid`
* Fix bug where templates were indexed by `FieldOrd`
* Fix incorrect SQL query syntax
* Fix broken creation/formatting of `.ki/*` files and gitignore
* Rewrite `gitIsDirty` to handle repos with no commits