Fix bug related to #114 where hashes file is not commited in `push`.
This commit fixes a bug where the working tree is left in a dirty state
after `write_collection()` has been called, at the end of a `ki push`
op. We have tests that `clone` and `pull` do not leave the working tree,
but I must have forgotten to write one for `push`.

* Add such a test, and make sure the file is committed.
* Bump version to `0.0.9a`.