Rewrite `htidy.c` and fix bugs in `htmlRegex, `guid` handling
Single forward slashes `/` are disallowed in this regex library, so they
must be escaped, which requires two preceding backslashes.

Our use of `showHex` threw a cryptic error about being passed a negative
integer (but there was no indication that it was `showHex` that was
throwing the error). This is fixed by using `Integer` instead of `Int`.

* Fix bug where trailing slash was omitted from basically all relative
  directory paths.
* Add `mkInternalDir` function.
* Add `castDir` function.
* Add `htmlInnerText()` function to `htidy.c`
* Completely rewrite `htidy.c`
* Fix git exception caused by empty media directory

This commit also adds a function `gitForceCommitAll`, which does not
check whether the root directory is dirty, and passes the
`--allow-empty` flag to `git commit`. This ensures that even if
`collection.media/` is empty, there is still at least one commit when we
perform our local clone to `_media/`.

This, in turn, means that the `git submodule add` invocation in the
target directory does not fail.