Fix bug where deck relative path was constructed wrong
This commit fixes a bug where the first component of each deck path was
at the end instead of the beginning (`b/c/d/e/a`). This happened because
we used a `foldr` call on a left-to-right ordered list, but initialized
the fold with the head of the list.

We also fix the `Show` instance for `ModelName`, and make `absify` only
work on `Dir`s. There was a bug in `absify` that omitted a trailing `/`
from constructed directory paths.