Decode protobuffer config fields with `proto-lens`
This commit implements boilerplate for getting the rest of the notetype
information into the YAML files dumped to the `_models/` directory.

I have tried several different protocol buffers libraries, and Google's
is the best, unsurprisingly. We are able to build the Haskell modules
from the `*.proto` files automatically at compile-time, and it gives us
nice proto-lens-specific types. In order to serialize the notetypes,
templates, and fields correctly, we will have to convert them to YAML.

Since the YAML library we are already using does this via `ToYAML`
instances, one way we can achieve this is by casting to a `Map`
representation stored in `FieldDef`, `Model`, and `Template` record
types. Then the YAML library will take care of the rest.

~~~~~~~~~~~~~~~~~
Technical details
~~~~~~~~~~~~~~~~~

We add `ToYAML` instances for `FieldDefConfig`, `NotetypeConfig`, and
`TemplateConfig`, which are themselves newtype wrappers around the
`proto-lens` types used to represent notetype config protobuffers.

This is a more concise alternative to a previous approach, which was
replicating all the data of the `proto-lens` types in new record types,
and writing smart constructors to convert between them.

We also use a custom schema encoder to serialize YAML scalars in such a
way that newline-containing strings are represented as blocks. This
makes it very nice to edit the CSS and LaTeX fields for notetypes.

Other minor changes:
* Rename `Git` -> `GIT` to match existing convention
* Don't warn about missing export lists to treat autogenerated source
  files
* Rename some functions to indicate that they work with `FieldDef`s
  instead of `Field`s
* Rename `getTemplate` -> `mkTemplate`
* Add `experiments/` to `.gitignore`
* Add dependency on `microlens`
* Remove a bunch of silly-looking extra newlines between top-level
  functions