Add scripts for reproducing submodule pull issue
Add working demo merge procedure in `repro-submodule-conflict.sh`

This version creates a new branch in the submodule, kills everything in
that branch, adds the changes from `remote`, and then merges that branch
into `master` within the submodule. The trouble is that git deletes new
files added on `master` because they are deleted in the new submodule
branch. So this doesn't really work.

Note that the file `d.txt` is missing from the submodule at the end