fix submodule.py script to handle empty directories
When converting a desk into a submodule, the original folder gets rm'ed from the git
repo. However, git rm does not always remove the folder: if there are untracked objects
inside the folder (such as an empty _media) folder, it does not remove it.

This commit adds a check for that, and if it was not removed manually removes the folder